IA-3365: add logging to profile bulk create
